+++ common/src/main/java/com/abrenoch/hyperiongrabber/common/HyperionNotification.java
@@ -12,18 +12,21 @@ import android.support.annotation.RequiresApi;
 import android.support.v4.app.NotificationCompat;
 
 public class HyperionNotification {
-    private static final String NOTIFICATION_CHANNEL_ID = "com.abrenoch.hyperiongrabber.notification";
-    private static final String NOTIFICATION_CHANNEL_LABEL = "Hyperion Grabber Notifications";
-    private static final String NOTIFICATION_TITLE = "Hyperion Grabber";
-    private static final String NOTIFICATION_DESCRIPTION = "Currently grabbing screen content";
+    private final String NOTIFICATION_CHANNEL_ID = "com.abrenoch.hyperiongrabber.notification";
+    private final String NOTIFICATION_CHANNEL_LABEL;
+    private final String NOTIFICATION_TITLE;
+    private final String NOTIFICATION_DESCRIPTION;
+    private final int PENDING_INTENT_REQUEST_CODE = 0;
     private final NotificationManager mNotificationManager;
     private final Context mContext;
     private Notification.Action mAction = null;
 
-
     HyperionNotification (Context ctx, NotificationManager manager) {
         mNotificationManager = manager;
         mContext = ctx;
+        NOTIFICATION_TITLE = mContext.getString(R.string.notification_title);
+        NOTIFICATION_DESCRIPTION = mContext.getString(R.string.notification_description);
+        NOTIFICATION_CHANNEL_LABEL = mContext.getString(R.string.notification_channel_label);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             mNotificationManager.createNotificationChannel(makeChannel());
         }
@@ -49,10 +52,17 @@ public class HyperionNotification {
     }
 
     public Notification buildNotification() {
+        Intent intent = mContext.getPackageManager().getLaunchIntentForPackage(mContext.getPackageName());
+        if (intent != null) {
+            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP);
+        }
+        PendingIntent pIntent = PendingIntent.getActivity(mContext, PENDING_INTENT_REQUEST_CODE, intent,
+                PendingIntent.FLAG_UPDATE_CURRENT);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             Notification.Builder builder = new Notification.Builder(mContext, NOTIFICATION_CHANNEL_ID)
                     .setOngoing(true)
                     .setSmallIcon(R.drawable.ic_notification_icon)
+                    .setContentIntent(pIntent)
                     .setContentTitle(NOTIFICATION_TITLE)
                     .setContentText(NOTIFICATION_DESCRIPTION);
             if (mAction != null) {
@@ -66,6 +76,7 @@ public class HyperionNotification {
                     .setOngoing(true)
                     .setPriority(Notification.PRIORITY_MAX)
                     .setSmallIcon(R.drawable.ic_notification_icon)
+                    .setContentIntent(pIntent)
                     .setContentTitle(NOTIFICATION_TITLE)
                     .setContentText(NOTIFICATION_DESCRIPTION);
             return builder.build();